requirements.txt
digi-xbee==1.3.0